rename package to kidspotz
