+++ app/src/main/java/com/example/kidspotz/LaunchMapsActivity.java
@@ -1,4 +1,4 @@
-package com.example.anushikhasharma.childsplay;
+package com.example.kidspotz;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
