+++ app/src/main/res/layout/activity_maps.xml
@@ -4,4 +4,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.mapwithmarker.MapsMarkerActivity" />
+    tools:context="com.example.kidspotz.MapsMarkerActivity" />
